feat(colors): add orange color support
Added support for the color orange in the printfcolor library. This includes updates to the color enumeration and ANSI escape codes for both foreground and background colors. The new color can be used in the printfc, printfc_fg, and printfc_bg functions. This enhancement allows for greater flexibility in text color customization for terminal output.